Initial Reliability calculation logic using octave
Given the MTBF of the product in hours, assuming the units are
used continuously 24x7 in the field.

It calculates how many units out a total number of units

(10, 50, 100, 1000, 5000, 10000 products)

in the field will fail for different periods of times like

1 year, 2, 3, 5, 6, 8, 10 years.

NOTE: Commiting after ages in 2018, so using a rough time when
the more scenarios based was used around that time. However The
basic logic I had done very much earlier.